Replace FileWatcher and JSONUtils with JSONLibraryUtils versions.
